Extract out legacy graph
