+++ packages/bundlers/default/src/DefaultBundler.js
@@ -1,9 +1,8 @@
 // @flow strict-local
+
 import type {
   Asset,
-  Bundle as LegacyBundle,
   BundleBehavior,
-  BundleGroup,
   Dependency,
   Environment,
   MutableBundleGraph,
@@ -21,24 +20,16 @@ import path from 'path';
 import {getFeatureFlag} from '@atlaspack/feature-flags';
 
 import {loadBundlerConfig, type ResolvedBundlerConfig} from './bundlerConfig';
+import {
+  decorateLegacyGraph,
+  type Bundle,
+  type DependencyBundleGraph,
+  type IdealGraph,
+} from './decorateLegacyGraph';
 import {addJSMonolithBundle} from './MonolithicBundler';
 
 /* BundleRoot - An asset that is the main entry of a Bundle. */
 type BundleRoot = Asset;
-export type Bundle = {|
-  uniqueKey: ?string,
-  assets: Set<Asset>,
-  internalizedAssets?: BitSet,
-  bundleBehavior?: ?BundleBehavior,
-  needsStableName: boolean,
-  mainEntryAsset: ?Asset,
-  size: number,
-  sourceBundles: Set<NodeId>,
-  target: Target,
-  env: Environment,
-  type: string,
-  manualSharedBundle: ?string, // for naming purposes
-|};
 
 const dependencyPriorityEdges = {
   sync: 1,
@@ -46,29 +37,6 @@ const dependencyPriorityEdges = {
   lazy: 3,
   conditional: 4,
 };
-
-type DependencyBundleGraph = ContentGraph<
-  | {|
-      value: Bundle,
-      type: 'bundle',
-    |}
-  | {|
-      value: Dependency,
-      type: 'dependency',
-    |},
-  number,
->;
-// IdealGraph is the structure we will pass to decorate,
-// which mutates the assetGraph into the bundleGraph we would
-// expect from default bundler
-type IdealGraph = {|
-  assets: Array<Asset>,
-  dependencyBundleGraph: DependencyBundleGraph,
-  bundleGraph: Graph<Bundle | 'root'>,
-  bundleGroupBundleIds: Set<NodeId>,
-  assetReference: DefaultMap<Asset, Array<[Dependency, Bundle]>>,
-  manualAssetToBundle: Map<Asset, NodeId>,
-|};
 
 /**
  *
@@ -125,213 +93,6 @@ export default (new Bundler({
   },
   optimize() {},
 }): Bundler);
-
-function decorateLegacyGraph(
-  idealGraph: IdealGraph,
-  bundleGraph: MutableBundleGraph,
-): void {
-  let idealBundleToLegacyBundle: Map<Bundle, LegacyBundle> = new Map();
-
-  let {
-    bundleGraph: idealBundleGraph,
-    dependencyBundleGraph,
-    bundleGroupBundleIds,
-    manualAssetToBundle,
-  } = idealGraph;
-  let entryBundleToBundleGroup: Map<NodeId, BundleGroup> = new Map();
-  // Step Create Bundles: Create bundle groups, bundles, and shared bundles and add assets to them
-  for (let [bundleNodeId, idealBundle] of idealBundleGraph.nodes.entries()) {
-    if (!idealBundle || idealBundle === 'root') continue;
-    let entryAsset = idealBundle.mainEntryAsset;
-    let bundleGroups = [];
-    let bundleGroup;
-    let bundle;
-
-    if (bundleGroupBundleIds.has(bundleNodeId)) {
-      invariant(
-        idealBundle.manualSharedBundle == null,
-        'Unstable Manual Shared Bundle feature is processing a manualSharedBundle as a BundleGroup',
-      );
-      let dependencies = dependencyBundleGraph
-        .getNodeIdsConnectedTo(
-          dependencyBundleGraph.getNodeIdByContentKey(String(bundleNodeId)),
-          ALL_EDGE_TYPES,
-        )
-        .map((nodeId) => {
-          let dependency = nullthrows(dependencyBundleGraph.getNode(nodeId));
-          invariant(dependency.type === 'dependency');
-          return dependency.value;
-        });
-      invariant(
-        entryAsset != null,
-        'Processing a bundleGroup with no entry asset',
-      );
-      for (let dependency of dependencies) {
-        bundleGroup = bundleGraph.createBundleGroup(
-          dependency,
-          idealBundle.target,
-        );
-        bundleGroups.push(bundleGroup);
-      }
-      invariant(bundleGroup);
-      entryBundleToBundleGroup.set(bundleNodeId, bundleGroup);
-
-      bundle = nullthrows(
-        bundleGraph.createBundle({
-          entryAsset: nullthrows(entryAsset),
-          needsStableName: idealBundle.needsStableName,
-          bundleBehavior: idealBundle.bundleBehavior,
-          target: idealBundle.target,
-          manualSharedBundle: idealBundle.manualSharedBundle,
-        }),
-      );
-
-      bundleGraph.addBundleToBundleGroup(bundle, bundleGroup);
-    } else if (
-      idealBundle.sourceBundles.size > 0 &&
-      !idealBundle.mainEntryAsset
-    ) {
-      let uniqueKey =
-        idealBundle.uniqueKey != null
-          ? idealBundle.uniqueKey
-          : [...idealBundle.assets].map((asset) => asset.id).join(',');
-
-      bundle = nullthrows(
-        bundleGraph.createBundle({
-          uniqueKey,
-          needsStableName: idealBundle.needsStableName,
-          bundleBehavior: idealBundle.bundleBehavior,
-          type: idealBundle.type,
-          target: idealBundle.target,
-          env: idealBundle.env,
-          manualSharedBundle: idealBundle.manualSharedBundle,
-        }),
-      );
-    } else if (idealBundle.uniqueKey != null) {
-      bundle = nullthrows(
-        bundleGraph.createBundle({
-          uniqueKey: idealBundle.uniqueKey,
-          needsStableName: idealBundle.needsStableName,
-          bundleBehavior: idealBundle.bundleBehavior,
-          type: idealBundle.type,
-          target: idealBundle.target,
-          env: idealBundle.env,
-          manualSharedBundle: idealBundle.manualSharedBundle,
-        }),
-      );
-    } else {
-      invariant(entryAsset != null);
-      bundle = nullthrows(
-        bundleGraph.createBundle({
-          entryAsset,
-          needsStableName: idealBundle.needsStableName,
-          bundleBehavior: idealBundle.bundleBehavior,
-          target: idealBundle.target,
-          manualSharedBundle: idealBundle.manualSharedBundle,
-        }),
-      );
-    }
-
-    idealBundleToLegacyBundle.set(idealBundle, bundle);
-
-    for (let asset of idealBundle.assets) {
-      bundleGraph.addAssetToBundle(asset, bundle);
-    }
-  }
-  // Step Internalization: Internalize dependencies for bundles
-  for (let idealBundle of idealBundleGraph.nodes) {
-    if (!idealBundle || idealBundle === 'root') continue;
-    let bundle = nullthrows(idealBundleToLegacyBundle.get(idealBundle));
-    if (idealBundle.internalizedAssets) {
-      idealBundle.internalizedAssets.forEach((internalized) => {
-        let incomingDeps = bundleGraph.getIncomingDependencies(
-          idealGraph.assets[internalized],
-        );
-        for (let incomingDep of incomingDeps) {
-          if (
-            incomingDep.priority === 'lazy' &&
-            incomingDep.specifierType !== 'url' &&
-            bundle.hasDependency(incomingDep)
-          ) {
-            bundleGraph.internalizeAsyncDependency(bundle, incomingDep);
-          }
-        }
-      });
-    }
-  }
-  // Unstable Manual Shared Bundles
-  // NOTE: This only works under the assumption that manual shared bundles would have
-  // always already been loaded before the bundle that requires internalization.
-  for (let manualSharedAsset of manualAssetToBundle.keys()) {
-    let incomingDeps = bundleGraph.getIncomingDependencies(manualSharedAsset);
-    for (let incomingDep of incomingDeps) {
-      if (
-        incomingDep.priority === 'lazy' &&
-        incomingDep.specifierType !== 'url'
-      ) {
-        let bundles = bundleGraph.getBundlesWithDependency(incomingDep);
-        for (let bundle of bundles) {
-          bundleGraph.internalizeAsyncDependency(bundle, incomingDep);
-        }
-      }
-    }
-  }
-
-  // Step Add to BundleGroups: Add bundles to their bundle groups
-  idealBundleGraph.traverse((nodeId, _, actions) => {
-    let node = idealBundleGraph.getNode(nodeId);
-    if (node === 'root') {
-      return;
-    }
-    actions.skipChildren();
-
-    let outboundNodeIds = idealBundleGraph.getNodeIdsConnectedFrom(nodeId);
-    let entryBundle = nullthrows(idealBundleGraph.getNode(nodeId));
-    invariant(entryBundle !== 'root');
-    let legacyEntryBundle = nullthrows(
-      idealBundleToLegacyBundle.get(entryBundle),
-    );
-
-    for (let id of outboundNodeIds) {
-      let siblingBundle = nullthrows(idealBundleGraph.getNode(id));
-      invariant(siblingBundle !== 'root');
-      let legacySiblingBundle = nullthrows(
-        idealBundleToLegacyBundle.get(siblingBundle),
-      );
-      bundleGraph.createBundleReference(legacyEntryBundle, legacySiblingBundle);
-    }
-  });
-
-  // Step References: Add references to all bundles
-  for (let [asset, references] of idealGraph.assetReference) {
-    for (let [dependency, bundle] of references) {
-      let legacyBundle = nullthrows(idealBundleToLegacyBundle.get(bundle));
-      bundleGraph.createAssetReference(dependency, asset, legacyBundle);
-    }
-  }
-
-  for (let {from, to} of idealBundleGraph.getAllEdges()) {
-    let sourceBundle = nullthrows(idealBundleGraph.getNode(from));
-    if (sourceBundle === 'root') {
-      continue;
-    }
-    invariant(sourceBundle !== 'root');
-
-    let legacySourceBundle = nullthrows(
-      idealBundleToLegacyBundle.get(sourceBundle),
-    );
-
-    let targetBundle = nullthrows(idealBundleGraph.getNode(to));
-    if (targetBundle === 'root') {
-      continue;
-    }
-    invariant(targetBundle !== 'root');
-    let legacyTargetBundle = nullthrows(
-      idealBundleToLegacyBundle.get(targetBundle),
-    );
-    bundleGraph.createBundleReference(legacySourceBundle, legacyTargetBundle);
-  }
-}
 
 function createIdealGraph(
   assetGraph: MutableBundleGraph,
